Fvanwij combat

+++ src/entity.c
@@ -76,9 +76,9 @@ void	update_enemy(t_entity *ent, t_app *cbd)
 	vec_normalize(&ent->dir);
 	// Determine if the entity is moving away from the player
 	if (vec_dot(ent->dir, vec_sub(cbd->playerdata.pos, ent->pos)) < 0)
-		ent->animation.current_animation = (9 - ent->health + 1);
+		ent->animation.current_animation = (9 - (int)ent->health + 1);
 	else
-		ent->animation.current_animation = (9 - ent->health);
+		ent->animation.current_animation = (9 - (int)ent->health);
 	if (ent->health == 0)
 		ent->animation.enabled = false;
 }
